+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v6.1.1 (2023-??-??)
 - fix: Multi-Month not included in fullcalendar-scheduler (#7177)
 - fix: Multi-Month has nonexistent 'internal' entrypoint, causing error for skypack (#7176)
 - fix: Vue connector should not error-out when given content-injector functions (#7175)
+- fix: continued CJS/ESM confusion with certain build tools (#7170, #7113, #7143)
 
 
 v6.1.0 (2023-01-29)